requirements.txt
Flask==3.0.0
numpy==1.26.0
pydub==0.25.1
Pillow==10.1.0
matplotlib==3.8.0
sounddevice==0.4.6
soundfile==0.12.1
scipy==1.11.3
opencv-python==4.8.1.78